// ui/src/ipc/generated/MergePreview.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MoneyView } from "./MoneyView";

export type MergePreview = { previewKey: string, fromLabel: string, intoLabel: string, fromTotal: MoneyView, intoTotal: MoneyView, total: MoneyView, };